backend/requirements.txt
fastapi
uvicorn[standard]
gurobipy
celery
redis
pydantic>=2.0
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
httpx
ruff
